test-adapter/pyproject.toml
[project]
name = "test-adapter"
version = "0.0.1"

dependencies = [
  "ai-edge-model-explorer"
]